Extract convertXmlAuditLogs test helper
