+++ packages/core/phase2/phase2.comparison.test.ts
@@ -5,7 +5,6 @@ import "../tests/helpers/setEnvironmentVariables"
 import type { AuditLogEvent } from "@moj-bichard7/common/types/AuditLogEvent"
 
 import { AuditLogEventSource } from "@moj-bichard7/common/types/AuditLogEvent"
-import { XMLParser } from "fast-xml-parser"
 import "jest-xml-matcher"
 
 import type { ParseIncomingMessageResult } from "../comparison/lib/parseIncomingMessage"
@@ -17,6 +16,7 @@ import CoreAuditLogger from "../lib/auditLog/CoreAuditLogger"
 import saveErrorListRecord from "../lib/database/saveErrorListRecord"
 import serialiseToXml from "../lib/serialise/pncUpdateDatasetXml/serialiseToXml"
 import checkDatabaseMatches from "../tests/helpers/comparison/checkDatabaseMatches"
+import convertXmlAuditLogs from "../tests/helpers/comparison/convertXmlAuditLogs"
 import {
   clearDatabase,
   disconnectDb,
@@ -102,36 +102,6 @@ const normaliseAuditLogs = (logs: AuditLogEvent[]): Partial<AuditLogEvent>[] =>
         ].includes(log.eventCode)
     )
 
-const convertXmlAuditLogs = (logs: string[]): AuditLogEvent[] => {
-  return logs.map((logXml) => {
-    const parser = new XMLParser()
-    const rawParsedObj = parser.parse(logXml)
-
-    const timestamp = rawParsedObj.logEvent.eventDateTime
-    const category = rawParsedObj.logEvent.eventCategory
-    const eventType = rawParsedObj.logEvent.eventType
-    const eventSource = rawParsedObj.logEvent.componentID
-    const attributes = rawParsedObj.logEvent.nameValuePairs.nameValuePair.reduce(
-      (acc: Record<string, string>, nvp: { name: string; value: string }) => {
-        acc[nvp.name] = nvp.value
-        if (nvp.name === "Force Owner") {
-          acc[nvp.name] = String(nvp.value).padStart(6, "0")
-        }
-
-        return acc
-      },
-      {}
-    )
-    const eventCode = attributes.eventCode
-    delete attributes.eventCode
-    const user = attributes.user ? { user: attributes.user } : {}
-    delete attributes.user
-    delete attributes.auditLogVersion
-
-    return { timestamp, category, eventCode, eventType, eventSource, attributes, ...user } as any as AuditLogEvent
-  })
-}
-
 describe("phase2", () => {
   beforeEach(async () => {
     await clearDatabase()
@@ -161,18 +131,21 @@ describe("phase2", () => {
     it("should generate the correct xml", () => {
       const expectedOutput = comparison.outgoingMessage
       const actualOutput = serialiseToXml(phase2Result.outputMessage, true)
+
       expect(actualOutput).toEqualXML(expectedOutput)
     })
 
     it("should generate the correct triggers", () => {
       const expectedTriggers = sortTriggers(comparison.triggers)
       const actualTriggers = sortTriggers(phase2Result.triggers)
+
       expect(actualTriggers).toStrictEqual(expectedTriggers)
     })
 
     it("should generate the correct exceptions", () => {
       const expectedExceptions = parsedOutgoingMessage.message.Exceptions
       const actualExceptions = phase2Result.outputMessage.Exceptions
+
       expect(actualExceptions).toStrictEqual(expectedExceptions)
     })
 
@@ -195,6 +168,7 @@ describe("phase2", () => {
     it("should generate the correct audit logs", () => {
       const expectedAuditLogs = normaliseAuditLogs(convertXmlAuditLogs(comparison.auditLogEvents))
       const actualAuditLogs = normaliseAuditLogs(phase2Result.auditLogEvents)
+
       expect(actualAuditLogs).toStrictEqual(expectedAuditLogs)
     })
   })
